Move server into packages
